Show a notification to let users know of the location of results
